new base: StochasticComponentBase. SourceBase inherits from that
